Add bazel build rules for the module.
* Split the part of the test that requires GCP into a separate e2e test
  because it needs to be run manually.
